KUDU-2717. Add an environment variable override for username
This adds a KUDU_USER_NAME environment variable which overrides the
lookup of the local unix user. This is a client-side identity which is
easy to spoof anyway (eg via LD_PRELOAD) so providing such an override
doesn't harm any security. But, it adds a certain amount of convenience
when trying to use kudu tools against an insecure cluster.

No tests are added since this is pretty trivial code, but I tested
manually that I was able to call 'kudu master set_flags' against a
remote cluster after specifying KUDU_USER_NAME=kudu

Change-Id: I9d42a3efbf26882f086046c276d5f9965a4068f5
Reviewed-on: http://gerrit.cloudera.org:8080/12627
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@apache.org>