[java] Add private diff scan support
Adds a private diff scan API to the java client. This is
achieved primarily by adding a
`diffScan(long startTimestamp, long endTimestamp)`
method to the AbstractKuduScannerBuilder builder,
and then configuring the scan RPC to include the set fields.

The column projection of the scan will include an extra
`IS_DELETED` column when a diff scan is performed.
The value of the `IS_DELETED` virtual column can be
retrieved via `RowResult.isDeleted()`

Change-Id: I51b01ae76cd22407df9097b56629ac7262ec2964
Reviewed-on: http://gerrit.cloudera.org:8080/12689
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Mike Percy <mpercy@apache.org>