KUDU-2741: De-flake TestMergeIterator.TestDeDupGhostRows
This test had a couple of test-only bugs related to bookkeeping,
resulting in this test being flaky in certain scenarios. This patch
fixes those bugs:

1. Ensure that we don't generate duplicate keys in the same list (a.k.a.
   RowSet) during the test because that can't occur "in the wild" and
   the MergeIterator does not support it.

2. Ensure that we don't maintain more than one instance of a single
   row key in the 'expected' result set. Switch to using an ordered map
   to enforce that test invariant. An additional benefit to using that
   data structure is that we can skip the sorting step when iterating
   the expected results.

Before this change, the test had a failure rate of about 10% according
to the flaky test dashboard. After this change, the tests passes
reliably and succeeded 1024/1024 times when I looped it using dist-test:

http://dist-test.cloudera.org/job?job_id=mpercy.1552437266.46341

Change-Id: I5ca9cdfa4620b1a9d4144de2289cc80ac37060aa
Reviewed-on: http://gerrit.cloudera.org:8080/12735
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>