[TSManager] don't hold lock while running LA command
This patch modifies the way how the location assignment is run
during the course of tablet server registration with master.

Prior to this patch, TSManager's lock on the registry of all
tablet servers was held while running the location assignment
command.  That might lead to stacking many TS registration
heartbeats in flight while master's service threads are waiting
to acquire the TSManager's lock.

This patch introduces changes to run the location assignment command
without holding the TSManager's lock.  That makes the registration
of tablet servers with masters more robust and makes the processing
of concurrent heartbeats from tablet servers more efficient overall.

Change-Id: I59cd5f6ed19c162a7c9f9a6527e78cab782b4539
Reviewed-on: http://gerrit.cloudera.org:8080/12749
Tested-by: Kudu Jenkins
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>