generic_iterators: switch MergeIterator to intrusive list of states
I originally wanted this for dominance because, in that world, a state is
either linked to the main list or to another state's dominated list. We're
going to go in a different direction, but an intrusive list still makes
sense because:
1. We don't need random access into states_, and
2. An intrusive list will yield O(1) erase without increasing the total
   number of allocations.

I also changed the MergeIterator tests to use int64_t instead of uint32_t
for values, as the latter reliably overflowed the non-overlapped input test
when num_lists >= 30.

Change-Id: I1a165858e37c3e0a6ef85e46e20078c264fa8a65
Reviewed-on: http://gerrit.cloudera.org:8080/12944
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Kudu Jenkins