[util] added cache with FIFO eviction policy
Updated cache implementation in util/cache.{h,cc} to accommodate
the notion of a cache with FIFO eviction policy. Such cache is
useful for backing a TTL cache, i.e. a cache where an entry's life
cycle depends purely on the time when the entry was first inserted
into the cache.

Corresponding unit test is added as well.

Change-Id: If06fcf6c9860d0f384816414205f8791cbe53480
Reviewed-on: http://gerrit.cloudera.org:8080/12777
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>