trace: fully qualify namespace in macros
Impala imports kudu/util/ but most of the Impala code is not in the
'kudu' namespace. In order for Impala to use the TRACE() macros, it's
thus necessary for them to fully qualify 'kudu::Trace' rather than
assume the macro is being used in a context where 'kudu::' is in scope.

Change-Id: I6365e08e85caa485611a8fdf2c329509911d49ab
Reviewed-on: http://gerrit.cloudera.org:8080/12834
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@apache.org>