tablet_server-test-base: make types match schema
This patch makes the types used in the TabletServerTestBase helper
methods consistent with the schema they assume, which specifies an int32
for both the key and the value fields.

Additionally, fix up the variable types in various places in
tablet_server-test to match the calls that are assigned to them.

Change-Id: I209db29d8be884db1173f5141f2e63a89c32d408
Reviewed-on: http://gerrit.cloudera.org:8080/12839
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins