util: helper class for working with bitsets
Adds a templatized helper class to facilitate working with bitsets. This
is just a wrapper around std::bitset, but it exposes a more
container-like interface. This is particularly useful for specifying
containers of enum types and the like (e.g. rather than using a hashed
container).

This also adds an iterator class for the new wrapper; such an iterator
apparently doesn't exist[1] in the STL bitset class.

[1] https://stackoverflow.com/a/34728458

Change-Id: Ie0344fe94e9f9da9651164cb1b456c92d99dbdf4
Reviewed-on: http://gerrit.cloudera.org:8080/12866
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>