authz: verify tokens on writes
Adds tserver-side privilege checking for write requests. Unlike the case
for scans, the privileges required for a given write request can only be
determined after the request is decoded, which currently happens in the
prepare phase of a write transaction. As such, this patch pushes
authorization into the prepare phase, right after decoding the
operations.

I also considered a couple of other approaches:
- Frontloading the decoding and doing the authorization at the front of
  the Write endpoint. This would need to take into consideration the
  taking of the schema lock when decoding, which currently happens on
  the prepare thread, which seems tricky.
- Decoding just the operation types up front and performing
  authorization at the front of the Write endpoint. I decided that the
  extra plumbing required to avoid doing two passes over the write
  request wouldn't be worth the added complexity.

Change-Id: Iefa2215d528a64f525e04bec111b25f8bc17c086
Reviewed-on: http://gerrit.cloudera.org:8080/11754
Tested-by: Kudu Jenkins
Reviewed-by: Hao Hao <hao.hao@cloudera.com>