java: fix tests that use ExpectedException
There's a weird interaction between the ExpectedException and
KuduTestHarness rules. The test passes, but the RetryRule embedded in
KuduTestHarness thinks it failed and reports a test failure. It has
something to do with the assumptions JUnit has around state; reinvoking a
test method violates those assumptions.

Anyway, the solution is to use a RuleChain to set up the two rules. In the
case of TestAsyncUtil, we didn't even need the ExpectedException at all.

Change-Id: Iddba2ccfac5a3120c111d25f5cfefae22292c5f4
Reviewed-on: http://gerrit.cloudera.org:8080/12875
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Kudu Jenkins