tablet_server_authorization-test: fix SIGSEGV with static linkage
There was a global Random instance that called LOG() when initialized.
That's a no-no as LOG() will crash if it is called before glog is
initialized. Sadly it wasn't caught by precommit as we force RELEASE builds
(which are normally statically linked) to use dynamic linkage.

Anyway, the fix is simple: make the Random instance a member of the test
fixture, and convert a few utility methods into fixture methods. The
'mutable' keyword is syntactic sugar that lets those methods remain 'const'.

Change-Id: I91d4988356a5bbd5e6b33ee08b7d49553be45227
Reviewed-on: http://gerrit.cloudera.org:8080/12899
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins