java/c++: ColumnSchema supports storing column comment
This patch intends to add a new api for storing column
comment in java client. And according to the comments
of the reviews, the empty comment is defined to delete
the comment for the uniform behavior.

In addition, the server-side has already been implemented
in: https://gerrit.cloudera.org/#/c/12849/

Change-Id: I4cdca4101898062cfe154c15ca40c5943d0e343c
Reviewed-on: http://gerrit.cloudera.org:8080/12890
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>