Enable tcmalloc heap sampling by default
This enables tcmalloc heap sampling by default with
`--heap-sample-every-n-bytes` set to 512KiB. This will allow users to
use the `pprof` tool to gather tcmalloc heap profiles from running Kudu
processes, at the cost of some overheard in memory allocation. I've been
doing a lot of performance testing on a cluster with heap sampling
enabled at this frequency and I don't think the overhead is a problem.
Enabling this soon after the release of 1.9 gives us lots of time to run
this configuration in tests to see if it causes any problems.

Change-Id: I99e06bdd04ecde030383709682cf096b5e5191c9
Reviewed-on: http://gerrit.cloudera.org:8080/12933
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Will Berkeley <wdberkeley@gmail.com>