[build] Fix building codegen on MacOS Mojave
When macOS_SDK_headers_for_macOS_10.14 is installed setting the
--sysroot to the default header location breaks the build, which is why
there was a check whether /usr/include existed and only set --sysroot to
the SDK path if it didn't.

The problem with this is if something creates the /usr/include
directory, the build will fail again. This commit changes the check to
look for /usr/include/libc.h instead.

Change-Id: I29ec74014bee3121bfb4a203dd7bfa8bac8f8321
Reviewed-on: http://gerrit.cloudera.org:8080/12936
Reviewed-by: Attila Bukor <abukor@apache.org>
Tested-by: Attila Bukor <abukor@apache.org>