KUDU-2449: "Tablet server addition" best practices
This section will detail the best practices around adding new tablet
servers to a cluster, specifically around node utilization. The most
common use-case for adding a tablet server is to reduce load/utilization
per node, or add additional capacity to the cluster. In either case, the
newly added tablet server instances won't be utilized unless new tablets are
created, or if tablets need to be re-homed, which is not the usual expectation
(I added new tablet servers, Kudu should detect that and use rebalance
automatically).

Change-Id: I7e44b4a84a8ca8048b8dc930cc1ee228daef6cf9
Reviewed-on: http://gerrit.cloudera.org:8080/12969
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@apache.org>