tools: escape brackets when generating XML
When building the site, we generate the XML for tools. Usually we escape
ambiguous characters for tooling arguments; it seems that wasn't the
case for arguments that are GFlags. This meant that before, we would run
into errors like the following when generating the site (new in 1.9.0
for the --predicates argument of the new `table scan` tool):

/kudu/build/release/gen-docs/kudu.xml:49: parser error : StartTag: invalid element name
 * The 'Comparison' type supports <=, <, =, >, and >=,
                                       ^

Change-Id: I16c13d86b0b452e0559e245ee33373078e5e3713
Reviewed-on: http://gerrit.cloudera.org:8080/12981
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Andrew Wong <awong@cloudera.com>
(cherry picked from commit 4eb74ae233fadf5184632f4b291b0befeb371dd5)
Reviewed-on: http://gerrit.cloudera.org:8080/12988
Tested-by: Kudu Jenkins