[mini-sentry] do not run the "store-cleaner" thread
In our MiniSentry-based tests we don't need to clean up the store since
the tests are run from a clean state anyway.  Not starting an extra
thread slightly improves startup times.  In my experiments about the
improvement was about ~100 milliseconds with the total startup time
to hover around 14 seconds.

In addition, this patch contains an update to always report on Sentry
startup times.  Prior to this patch, that report appeared only if
it took more than half of the configured timeout interval for Sentry's
startup.

Change-Id: I908d00d93a9f4cf632899da863275dd01b4a2fad
Reviewed-on: http://gerrit.cloudera.org:8080/12995
Tested-by: Kudu Jenkins
Reviewed-by: Hao Hao <hao.hao@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>