KUDU-2762: Improve write transaction tracing
The trace logging for write transactions had grammatical differences as
well as errors in punctuation. In addition, the `APPLY` phase tracing
was missing an end message while the `REPLICATION` phase tracing was
missing both a start and end message. Finally, trace messaging within
tablet.cc was prepending the `PREPARE` phase to it unnecessarily.

Change-Id: I3e69e45210c1c284d9e43c4e47fe052c41a353d7
Reviewed-on: http://gerrit.cloudera.org:8080/12999
Tested-by: Kudu Jenkins
Reviewed-by: Attila Bukor <abukor@apache.org>
Reviewed-by: Andrew Wong <awong@cloudera.com>