generic_iterators: refactor MergeIterator for whole-block-copy optimization
This patch refactors MergeIterator::MaterializeBlock in preparation for the
whole-block-copy optimization. Specifically, the "copy next row" and
"advance iter and rebalance heaps" operations have been decomposed into
standalone functions so that it'll be easier to add "copy whole block" as
an equivalent function in the future.

Also of note: we no longer set the entire client selection vector up front,
as this won't necessarily be the case when we copy a block.

Change-Id: I050116edc51bb3e91852e6286c221175770e6382
Reviewed-on: http://gerrit.cloudera.org:8080/13010
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Adar Dembo <adar@cloudera.com>