SelectionVector: pad extra bits with zeroes in constructor
I found this after removing the MergeIterator's call to SetAllTrue(). It
turns out that if you don't call any functions that set all bytes en masse,
CountSelected() and AnySelected() will misbehave as they'll read garbage
data from any trailing bits beyond the logical end of the bitmap.

We can fix this in one of two ways:
- Modify CountSelected()/AnySelected() to ignore the trailing bits.
- Zero out the trailing bits in the constructor.

I opted for the second approach as I found it easier to implement, and I
suspect it's more performant than the first.

Change-Id: I74baf87c7f38b7eea0ad46825c2421c99f8f42a1
Reviewed-on: http://gerrit.cloudera.org:8080/13009
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>