[build] Fix the java compatibility checker script
This patch fixes check_compatibility.py. Primarily it
changed the existing script to use the Gradle build
in place of the Maven build that no longer exists.

It now uses japicmp in place of java_acc because
java_acc had issues unzipping the jars created by the
kudu build. Additionally japicmp’s filtering seemed to
work better.

The patch also updates the Gradle build to ensure
generated Protobuf classes are annotated with
@Generated to make report filtering easier.

In follow on patches I may incorporate this into the LINT job
and cause failure on API breakage.

Change-Id: Id482da0b3aa5ea35fcfb2b674f0d7ae413045ca3
Reviewed-on: http://gerrit.cloudera.org:8080/13004
Tested-by: Kudu Jenkins
Reviewed-by: Greg Solovyev <gsolovyev@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>