dist-test: support test result reporting
The goal is to be able to use dist-test in any run of build-and-test.sh,
even those that power the flaky test dashboard.

Because dist-test runs just one test per invocation, it isn't safe to disown
the reporting process; the slave might disappear mid-report. There wasn't an
obvious way to know that we're using dist-test from within run-test.sh, so I
opted to stop disowning altogether. The Java result reporting is synchronous
(and per-test rather than per-class) so this didn't seem like a big deal.

I was concerned that the additional parallelism might lead to "thundering
herd" behavior in the test result server (i.e. if all tests finished and
reported at the same time) so I ran a couple reporting-enabled dist-tests
looking for load spikes on the server. I didn't see any, likely because:
- Most tests' running time varies.
- Tests only start when dist-test slaves become available, not all at once.

RUN_FLAKY_ONLY will be addressed in a follow-up.

Change-Id: Ie5072e2395376e455f63dac6d3debc88526aed07
Reviewed-on: http://gerrit.cloudera.org:8080/13068
Tested-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>