dist-test: support RUN_FLAKY_ONLY
The flaky test list is somewhat short so there's no real reason for the
tests to be distributed. However, over the years we've observed that the
dist-test runtime environment is different enough from non-dist-test
environments that a test may be flaky in one but not the other. Since the
purpose of flaky test resistance is to improve the precommit experience, and
since all precommit tests run in dist-test, the test runs that power the
flaky test dashboard (i.e. RUN_FLAKY_ONLY=1) should follow suit and also run
in dist-test.

Change-Id: Id5a063512f82341a48031911de12c44f902a2723
Reviewed-on: http://gerrit.cloudera.org:8080/13072
Tested-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>