[backup] Quick fix to fromMs option logic
The fromMS logic was incorrectly updated and we
didn’t have a test validating the bahavior. This patch
fixes and tests that issue.

Additionally the TestForceIncrementalBackup was
wrong as well.

KUDU-2790 tracks overall coverage improvements,
so this is just a quick fix. KUDU-2790  will be my next
path on this code.

Change-Id: I12dc557fa71470ca8aef2a1c2393bda28e0b81dd
Reviewed-on: http://gerrit.cloudera.org:8080/13105
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins