hms: remove experimental tag from hms flags
The HMS integration has been in our codebase for a while now. While the
behavior of integration may change (a la KUDU-2799), the configuration
options themselves aren't likely to do so. This patch tags the necessary
configuration flags as "evolving" (the default) to facilitate their
usage and testing of the integration.

Change-Id: I4f862a805a370c250a712347d48965ed71fbd78e
Reviewed-on: http://gerrit.cloudera.org:8080/13124
Tested-by: Kudu Jenkins
Reviewed-by: Hao Hao <hao.hao@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>