[util] modernize signature of Cache interface methods
Updated the signature of the Lookup() and Insert() methods of the Cache
interface to return UniqueHandle.  That helps in automating reference
counting for the cached items.

Also, Release() and Free() methods have been moved into the protected
section.  With UniqueHandle returned from Lookup() and Insert(),
Release() and Free() are no longer required to be the public part
of the Cache's interface.

Change-Id: I31518cdfa269e6806aa5dda180956896d8cb3404
Reviewed-on: http://gerrit.cloudera.org:8080/13175
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Alexey Serbin <aserbin@cloudera.com>