[backup] Support partition alterations between Kudu backups
This patch enables partitions to be added and dropped
between backups.

It does this by adding tablet and partition details to the
metadata which can be used to detect if a partition
was dropped. It filters out partitions that are no longer
valid and adds the remaining partitions to a
KuduPartitioner. That KuduPartitioner can then be used
to filter out non-covered rows.

This does require some previously package/private
classes and constructors to be marked as public.
In those cases I have tagged them as Private and Unstable
using the interface annotations.

Change-Id: I31e0eb27f163c38840e5466ff85d0b4a44d4ec0a
Reviewed-on: http://gerrit.cloudera.org:8080/13191
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Grant Henke <granthenke@apache.org>