[util] interface to invalidate cache entries
Added a means to invalidate entries in Cache.  Basically, it's a way
to iterate over all entries in a cache while having an ability to
remove an entry based on a set of criteria specified using an entry's
key and value.  The generic usage pattern of the introduced invalidation
mechanism is to purge of entries matching criteria based on their keys
and values.

In particular, that's useful for periodic scrubbing tasks which get
rid of expired entries in a TTLCache.  It's about invalidating expired
entries in a TTL cache even if the cache hasn't reached its capacity
yet.  The latter, being an optimization, allows for more effective
usage of the OS memory in the presence of multiple concurrent processes.

Added corresponding tests as well.

Change-Id: I1e340c3270b5580114151cdd40afe82576c18099
Reviewed-on: http://gerrit.cloudera.org:8080/13196
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>